move OapUtils.getOutputPathBasedonConf to IndexUtils
