+++ src/main/scala/org/apache/spark/sql/execution/datasources/oap/index/IndexUtils.scala
@@ -22,7 +22,9 @@ import java.io.OutputStream
 import org.apache.hadoop.conf.Configuration
 import org.apache.hadoop.fs.Path
 
+import org.apache.spark.sql.RuntimeConfig
 import org.apache.spark.sql.catalyst.InternalRow
+import org.apache.spark.sql.execution.datasources.FileIndex
 import org.apache.spark.sql.execution.datasources.oap.OapFileFormat
 import org.apache.spark.sql.execution.datasources.oap.io.{BytesCompressor, BytesDecompressor, IndexFile}
 import org.apache.spark.sql.internal.oap.OapConf
@@ -104,10 +106,11 @@ private[oap] object IndexUtils {
     if (indexDirectory != "") {
       new Path(
         indexDirectory + "/" + Path.getPathWithoutSchemeAndAuthority(dataFile.getParent),
-        "." + indexFileName + "." + time + "." + name + OAP_INDEX_EXTENSION)
+        s"${"." + indexFileName + "." + time + "." + name + OAP_INDEX_EXTENSION}")
     } else {
       new Path(
-        dataFile.getParent, "." + indexFileName + "." + time + "." + name + OAP_INDEX_EXTENSION)
+        dataFile.getParent,
+        s"${"." + indexFileName + "." + time + "." + name + OAP_INDEX_EXTENSION}")
     }
   }
 
@@ -190,6 +193,35 @@ private[oap] object IndexUtils {
         "." + indexFileName + extension)
     }
   }
+
+  /**
+   * Generate the outPutPath based on OapConf.OAP_INDEX_DIRECTORY and the data path,
+   * here the dataPath does not contain the partition path
+   * @param fileIndex [[FileIndex]] of a relation
+   * @param conf the configuration to get the value of OapConf.OAP_INDEX_DIRECTORY
+   * @return the outPutPath to save the job temporary data
+   */
+  def getOutputPathBasedOnConf(fileIndex: FileIndex, conf: RuntimeConfig): Path = {
+    def getTableBaseDir(path: Path, times: Int): Path = {
+      if (times > 0) getTableBaseDir(path.getParent, times - 1)
+      else path
+    }
+    val paths = fileIndex.rootPaths
+    assert(paths.nonEmpty, "Expected at least one path of fileIndex.rootPaths, but no value")
+    val dataPath = paths.length match {
+      case 1 => paths.head
+      case _ => getTableBaseDir(paths.head, fileIndex.partitionSchema.length)
+    }
+
+    val indexDirectory = conf.get(OapConf.OAP_INDEX_DIRECTORY.key)
+    if (indexDirectory != "") {
+      new Path (
+        indexDirectory + Path.getPathWithoutSchemeAndAuthority(dataPath).toString)
+    } else {
+      dataPath
+    }
+  }
+
   val INT_SIZE = 4
   val LONG_SIZE = 8
