+++ src/main/scala/org/apache/spark/sql/execution/datasources/oap/index/indexPlans.scala
@@ -159,7 +159,7 @@ case class CreateIndexCommand(
     }
     // Generate the outPutPath based on OapConf.OAP_INDEX_DIRECTORY and the data path,
     // here the dataPath does not contain the partition path
-    val outPutPath = OapUtils.getOutputPathBasedConfiguration(fileCatalog, sparkSession.conf)
+    val outPutPath = IndexUtils.getOutputPathBasedOnConf(fileCatalog, sparkSession.conf)
     assert(outPutPath != null, "Expected exactly one path to be specified, but no value")
 
     val qualifiedOutputPath = {
@@ -397,7 +397,7 @@ case class RefreshIndexCommand(
 
       // Generate the outPutPath based on OapConf.OAP_INDEX_DIRECTORY and the data path,
       // here the dataPath does not contain the partition path
-      val outPutPath = OapUtils.getOutputPathBasedConfiguration(fileCatalog, sparkSession.conf)
+      val outPutPath = IndexUtils.getOutputPathBasedOnConf(fileCatalog, sparkSession.conf)
       assert(outPutPath != null, "Expected exactly one path to be specified, but no value")
 
       val qualifiedOutputPath = {
